US121 Reorganize CompletableRow and remove Menu
Removed an unused menu and reorganized the CompletableRow into its own
folder along with components that it only uses.
