+++ client/src/main/components/CompletableRow/CompletedCheckbox.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Checkbox } from '@material-ui/core';
-import { Project, Task } from '../logic/dbTypes';
+import { Project, Task } from '../../logic/dbTypes';
 
 export type CompletedCheckboxProps = {
   className?: string;
